refactor(cli): move the sniff command group out of core/cli.py
Extract the `sniff` group (ble, zigbee, thread, lora, airtag_scanner) into
the new `modules/sniff` package. The command bodies are moved verbatim; only
the import depth changes. `core/cli.py` goes from 1472 to 1019 lines.

With `sniff` gone, `core/cli.py` no longer depends on `core/extcap.py` at
all -- `find_putty_path` and `run_extcap_directly` were its only users. The
imports left unused by the move are dropped as well: `run_bridge`,
`run_sx_bridge`, `SniffingFirmware`, `SniffingBaseFirmware`, `Catnip` and
`CatSnifferDevice` (the last one already unused before this phase).

`modules/sniff/__init__.py` is kept empty on purpose: `core/cli.py` imports
it statically, so anything eager there would be paid on every `catnip --help`.

Verified: no diff against tests/snapshots/cli_tree_linux.txt, `find_packages`
still lists the new package, and `pytest tests/ -q` gives 328 passed / 3
failed -- identical to a worktree at HEAD.
