+++ catnip/modules/sniff/__init__.py
@@ -1,0 +1,5 @@
+"""``catnip sniff`` - per-protocol sniffing commands.
+
+Deliberately empty: ``modules/core/cli.py`` imports ``modules.sniff.cli``
+statically, so anything eager here would be paid on every ``catnip --help``.
+"""
